EDC_verifier_1/requirements.txt
pandas==0.25.3
paho-mqtt
pycryptodome
